requirements.txt
zenml[server]==0.30
numpy==1.23.5
google-cloud-bigquery==3.4.1
pandas==1.5.2
pandas-gbq==0.18.1
scikit-learn==1.2.0
xgboost==1.7.2
imbalanced-learn==0.10.1
plotly==5.11.0
db-dtypes==1.0.5
matplotlib==3.6.2
mlflow==2.1.1
bentoml==1.0.12
